backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.10.0
twilio==9.4.0
httpx==0.27.2
pydantic==2.10.3
websockets==13.1
openai==1.57.4
elevenlabs==1.50.4
